backend_django/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.11
django-extensions==3.2.3
django-cache-url==3.4.5
django-debug-toolbar==3.2.1
django-phonenumber-field==7.3.0
environs==9.3.2
marshmallow==3.20.2
packaging==23.2
Pillow==10.3.0
python-dotenv==1.0.1
pytz==2024.1
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
charset-normalizer==3.3.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
geopy==2.4.1
requests==2.31.0
idna==3.6
geographiclib==2.0
phonenumbers==8.13.31
urllib3==2.2.1
GitPython==3.1.43
rollbar==1.0.0
gevent==24.2.1
gunicorn==23.0.0
